updating todo

+++ src/main/java/com/mavha/mavhachallengetodobespring/domain/Todo.java
@@ -2,8 +2,6 @@ package com.mavha.mavhachallengetodobespring.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -20,7 +18,7 @@ public class Todo {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	// TODO: rename to title
 	@Column(name = "name", nullable = false)
